Updated observability.py
- Moved duckdb UI based methods to its own class inside observability module
